+++ package/environments/production.ts
@@ -80,4 +80,4 @@ Content hashes get added to the filenames regardless of setting useContentHash i
 `)
 }
 
-module.exports = merge(baseConfig, productionConfig)
+export default merge(baseConfig, productionConfig)